backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.13.5",
    "chromadb>=1.0.20",
    "crawl4ai>=0.7.6",
    "fastapi>=0.116.1",
    "langchain>=0.3.27",
    "langchain-chroma>=0.2.5",
    "langchain-community>=0.3.27",
    "langchain-google-genai>=2.1.10",
    "langchain-openai>=0.3.30",
    "pypdf>=6.0.0",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
    "sentence-transformers>=5.1.1",
    "uvicorn>=0.35.0",
]
